Rohit/agentic_ai_testing/outputs/01_requirement_analysis.txt

Feature: Login Functionality Testing

Test Scenarios:
1. Valid User Login
   - User enters valid username and password
   - Expected: User successfully logs in and redirected to dashboard
   
2. Invalid Username
   - User enters invalid username
   - Expected: Error message displayed, user remains on login page
   
3. Invalid Password
   - User enters valid username but wrong password
   - Expected: Error message displayed, user remains on login page
   
4. Empty Fields
   - User leaves username or password empty
   - Expected: Validation error message displayed
   
5. Case Sensitivity
   - Test if username/password are case-sensitive
   - Expected: Should handle appropriately based on requirement
